refactor: redis repository 리펙터링
